revert link changes
