+++ src/app/home/components/Card/DashboardCard.tsx
@@ -1,6 +1,6 @@
 /** @jsx jsx */
 import { jsx } from '@emotion/core';
-import React, { Component } from 'react';
+import React, { useState } from 'react';
 import {
   Card,
   CardHeader,
@@ -14,7 +14,7 @@ import theme from '../../../../theme';
 import Loader from 'react-loader-spinner';
 import CardStatus from './Status/CardStatus';
 import MigrationStatus from './Status/MigrationStatus';
-import FooterText from './FooterText';
+import CardFooterText from './CardFooterText';
 import HeaderText from './HeaderText';
 import { css } from '@emotion/core';
 import { Flex, Box, Text } from '@rebass/emotion';
@@ -33,76 +33,68 @@ interface IProps {
   expandDetails?: (string) => void;
 }
 
-class DashboardCard extends Component<IProps, IState> {
-  state = {
-    isOpen: false,
-  };
+const DashboardCard: React.FunctionComponent<IProps> = (
+  {
+    dataList,
+    isFetching,
+    type,
+    isError,
+    planStatusCounts,
+    expandDetails
+  }
+) => {
+  const [isOpen, toggleOpen] = useState(false);
 
-  onToggle = isOpen => {
-    this.setState({
-      isOpen,
-    });
-  };
-
-  onSelect = event => {
-    this.setState({
-      isOpen: !this.state.isOpen,
-    });
-  };
-  render() {
-    const { dataList, title, isFetching, type, isError, planStatusCounts, expandDetails } = this.props;
-    const { isOpen } = this.state;
-    if (isError) {
-      return (
-        <Card style={{ minHeight: '100%', height: '16em' }}>
-          <div className="pf-l-flex">
-            <div className="pf-l-flex__item">
-              <StatusIcon isReady={false} />
-            </div>
-            <div className="pf-l-flex__item">
-              Failed to fetch
-            </div>
-          </div>
-        </Card>
-      );
-    }
+  if (isError) {
     return (
       <Card style={{ minHeight: '100%', height: '16em' }}>
-        {dataList && !isFetching ? (
-          <React.Fragment>
-            <CardHeader>
-              <Title size="md">
-                <HeaderText type={type} dataList={dataList} />
-              </Title>
-            </CardHeader>
-            <CardBody>
-              {type === 'plans' ? (
-                <MigrationStatus planStatusCounts={planStatusCounts} />
-              ) : (
-                  <CardStatus dataList={dataList} type={type} />
-                )}
-            </CardBody>
-            <CardFooter>
-              <FooterText dataList={dataList} type={type} expandDetails={expandDetails} />
-            </CardFooter>
-          </React.Fragment>
-        ) : (
-            <Flex
-              css={css`
+        <div className="pf-l-flex">
+          <div className="pf-l-flex__item">
+            <StatusIcon isReady={false} />
+          </div>
+          <div className="pf-l-flex__item">
+            Failed to fetch
+            </div>
+        </div>
+      </Card>
+    );
+  }
+  return (
+    <Card style={{ minHeight: '100%', height: '16em' }}>
+      {dataList && !isFetching ? (
+        <React.Fragment>
+          <CardHeader>
+            <Title size="md">
+              <HeaderText type={type} dataList={dataList} />
+            </Title>
+          </CardHeader>
+          <CardBody>
+            {type === 'plans' ? (
+              <MigrationStatus planStatusCounts={planStatusCounts} />
+            ) : (
+                <CardStatus dataList={dataList} type={type} />
+              )}
+          </CardBody>
+          <CardFooter>
+            <CardFooterText dataList={dataList} type={type} expandDetails={expandDetails} />
+          </CardFooter>
+        </React.Fragment>
+      ) : (
+          <Flex
+            css={css`
               height: 100%;
               text-align: center;
               margin: auto;
             `}
-            >
-              <Box flex="1" m="auto">
-                <Loader type="ThreeDots" color={theme.colors.navy} height="100" width="100" />
-                <Text fontSize={[2, 3, 4]}> Loading </Text>
-              </Box>
-            </Flex>
-          )}
-      </Card>
-    );
-  }
+          >
+            <Box flex="1" m="auto">
+              <Loader type="ThreeDots" color={theme.colors.navy} height="100" width="100" />
+              <Text fontSize={[2, 3, 4]}> Loading </Text>
+            </Box>
+          </Flex>
+        )}
+    </Card>
+  );
 }
 
 export default DashboardCard;
